# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Video Streaming Service API
  version: 1.0.0
  contact:
    name: video-streaming-service
    url: https://github.com/Nuu-maan/video-streaming-service
  description: |
    Standalone JSON API for a video platform: accounts, uploads, transcoding,
    HLS/MP4 playback, comments, playlists, subscriptions, search, moderation
    and admin analytics. It is designed to be hosted on its own and consumed
    by a separate frontend (React, Next.js, Vue, ...) running on a different
    origin.

    ## The three things to know first

    1. **Auth is JWT Bearer.** Log in at `POST /auth/login` (the body field is
       `identifier` — a username *or* an email — plus `password`) and send
       `Authorization: Bearer <access_token>` on every authenticated call.
       Access tokens live 900 s; exchange the `refresh_token` at
       `POST /auth/refresh` for a new pair. A refresh token is *not* usable as
       an API credential, and an access token is rejected at `/auth/refresh`.

    2. **Every JSON response is wrapped in an envelope.**
       Success: `{"success": true, "data": ...}`.
       Lists (including `/search`): `{"success": true, "data": [...], "pagination": {"total", "page", "limit", "total_pages", "has_next", "has_previous"}}`.
       Errors: `{"success": false, "error": {"code": "...", "message": "..."}}`.
       Media routes (`.m3u8`, `.ts`, MP4, JPEG) return raw bytes, never the envelope.

...

    ## Other conventions

...
servers:
...
    variables:
...
security: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      operationId: register
      summary: Create an account and return tokens
      description: >-
        Registers a new account and immediately returns a token pair, so no
        separate login round-trip is needed. Usernames must match
        `^[a-zA-Z0-9_]{3,30}$`.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                  minLength: 3
                  maxLength: 30
                  pattern: "^[a-zA-Z0-9_]{3,30}$"
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          description: The account is banned (`USER_BANNED`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username or email already taken (`ALREADY_EXISTS`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Exchange credentials for tokens
      description: >-
        **The credential field is `identifier`, not `username`** — it accepts a
        username or an email. Unknown account and wrong password return the
        same 401 body, so account existence cannot be probed.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier:
                  type: string
                  description: Username **or** email
                password:
                  type: string
      responses:
        "200":
          description: Token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: The account is banned (`USER_BANNED`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshToken
      summary: Exchange a refresh token for a new token pair
      description: >-
        Takes `{"refresh_token": "..."}` and returns a full new token pair.
        An **access** token presented here is rejected with 401, just as a
        **refresh** token presented as an API credential is rejected with 401.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPairResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          description: Invalid, expired or revoked refresh token — or an access token was presented
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/me:
    get:
      tags: [Auth]
      operationId: getMe
      summary: Return the authenticated caller's own account
      security:
        - bearerAuth: []
      responses:
        "200":
          description: The caller's account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Revoke the presented access token
      description: >-
        Revokes the access token in the `Authorization` header. The body is
        optional; when it carries a `refresh_token`, that token is revoked too.
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/logout-all:
    post:
      tags: [Auth]
      operationId: logoutAll
      summary: Revoke every outstanding session for the caller, on every device
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All sessions revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ─────────────────────────── Account ───────────────────────────
  /auth/verify-email/send:
    post:
      tags: [Account]
      operationId: sendVerificationEmail
      summary: (Re)send a verification email
      description: Sends the verification mail to the caller's own registered address.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Verification email sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Address already verified (`EMAIL_ALREADY_VERIFIED`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/verify-email:
    post:
      tags: [Account]
      operationId: verifyEmail
      summary: Consume a verification token and mark the account verified
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
      responses:
        "200":
          description: Email verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad or expired token (`INVALID_TOKEN`) or malformed body (`VALIDATION_ERROR`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/forgot-password:
    post:
      tags: [Account]
      operationId: forgotPassword
      summary: Start a password reset
      description: >-
        Always answers 200 with the same body whether or not the address is
        registered, so accounts cannot be enumerated.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  format: email
      responses:
        "200":
          description: Accepted (same body regardless of whether the address exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
  /auth/reset-password:
    post:
      tags: [Account]
      operationId: resetPassword
      summary: Consume a reset token and set a new password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Password reset; log in again with the new password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Bad token (`INVALID_TOKEN`) or weak password (`VALIDATION_ERROR`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /me/change-password:
    post:
      tags: [Account]
      operationId: changePassword
      summary: Change password after verifying the current one
      description: Carries the stricter auth rate limit on top of the normal API budget.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                new_password:
                  type: string
      responses:
        "200":
          description: Password changed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Wrong current password (`INVALID_CURRENT_PASSWORD`) or weak new password (`VALIDATION_ERROR`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─────────────────────────── Videos ───────────────────────────
  /videos:
    get:
      tags: [Videos]
      operationId: listVideos
      summary: List videos
      description: >-
        Auth is optional. Anonymous callers see only public videos; with a
        token and `mine=true` the filter switches to the caller's own videos,
        all visibilities included.
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Limit"
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/VideoStatus"
        - name: mine
          in: query
          description: '`"true"` lists the caller''s own videos (requires a token)'
          schema:
            type: string
            enum: ["true"]
      responses:
        "200":
          description: Page of videos
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Video"
        "500":
          description: Internal error (`INTERNAL_ERROR`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /videos/upload:
    post:
      tags: [Videos]
      operationId: uploadVideo
      summary: Upload a video for transcoding
      description: >-
        Multipart upload, queued for transcoding by the worker. Requires the
        `upload_video` permission and carries an extra upload rate limit.
        Title 1–255 chars. The size limit is deployment-configured
        (`STORAGE_MAX_FILE_SIZE`, default 2 GiB; hard ceiling 10 GiB). The
        response video starts in status `uploading` — poll
        `GET /videos/{id}/status` for progress.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [video, title]
              properties:
                video:
                  type: string
                  format: binary
                  description: The video file
                title:
                  type: string
                  minLength: 1
                  maxLength: 255
                description:
                  type: string
                visibility:
                  $ref: "#/components/schemas/VideoVisibility"
      responses:
        "201":
          description: Video accepted (status `uploading`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "413":
          description: File exceeds the configured size limit (`FILE_TOO_LARGE`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "415":
          description: Not an accepted video format (`INVALID_FORMAT`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /videos/{id}:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    get:
      tags: [Videos]
      operationId: getVideo
      summary: Get one video
      description: >-
        Auth is optional; a token lets the owner read their own private video.
        **Private videos answer 404 — never 403 — to everyone else**, so their
        existence is not leaked. Unlisted videos resolve for anyone with the
        link. A ready video carries `hls_url` and `thumbnail_url` for playback.
      responses:
        "200":
          description: The video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Videos]
      operationId: deleteVideo
      summary: Delete a video
      description: >-
        Owner only, unless the caller holds `delete_any_video` (checked in the
        handler).
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Video deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /videos/{id}/status:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    get:
      tags: [Videos]
      operationId: getVideoStatus
      summary: Transcoding progress for a video
      description: >-
        Poll this after an upload. Auth is optional; private videos 404 for
        non-owners exactly like `GET /videos/{id}`.
      responses:
        "200":
          description: Processing status
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/VideoStatusReport"
        "400":
          $ref: "#/components/responses/ValidationError"
        "404":
          $ref: "#/components/responses/NotFound"

  # ─────────────────────────── Engagement ───────────────────────────
  /videos/{id}/view:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    post:
      tags: [Engagement]
      operationId: recordView
      summary: Record one view (explicit — playback does not auto-count)
      description: >-
        Views are deduplicated in Redis. Auth is optional: authenticated
        callers are deduped by user id; **anonymous callers must send a
        `session_id`**. The body must be JSON, even if just `{}` for an
        authenticated caller. **201 = counted, 200 = deduplicated repeat.**
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                quality:
                  type: string
                source:
                  type: string
                session_id:
                  type: string
                  description: Required for anonymous callers
      responses:
        "201":
          description: View counted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ViewResult"
        "200":
          description: Duplicate within the dedupe window — not counted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ViewResult"
        "400":
          description: Missing `session_id` for an anonymous caller, or non-JSON body (`VALIDATION_ERROR`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /me/notifications/unread-count:
    get:
      tags: [Social]
      operationId: unreadNotificationCount
      summary: Unread notification count for badge rendering
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Unread count
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          unread_count:
                            type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        `Authorization: Bearer <access_token>`. Obtain a token pair from
        `/auth/login` or `/auth/register`; renew with `/auth/refresh`.
        Access tokens live 900 s. A refresh token is not accepted here.
...